Adding array evaluation.
Evaluation of the array is done by converting an evaluated array
into a function interpretation, then using the function interpretation
extraction mechanism.

The example has been updated to contain a small use of each. Although
not a full practical example, it can serve as a simple test and showing
how they can be used.